Fix variant-limiting bug in matdbg

+++ filament/src/Material.cpp
@@ -479,9 +479,9 @@ void FMaterial::onEditCallback(void* userdata, const utils::CString& name, const
             packageSize);
 }
 
-void FMaterial::onQueryCallback(void* userdata, uint16_t* pvariants) {
+void FMaterial::onQueryCallback(void* userdata, uint64_t* pvariants) {
     FMaterial* material = upcast((Material*) userdata);
-    uint16_t variants = 0;
+    uint64_t variants = 0;
     auto& cachedPrograms = material->mCachedPrograms;
     for (size_t i = 0, n = cachedPrograms.size(); i < n; ++i) {
         if (cachedPrograms[i]) {
