+++ libs/matdbg/src/DebugServer.cpp
@@ -471,7 +471,7 @@ const DebugServer::MaterialRecord* DebugServer::getRecord(const MaterialKey& key
 void DebugServer::updateActiveVariants() {
     if (mQueryCallback) {
         for (auto& pair : mMaterialRecords) {
-            uint16_t& result = mMaterialRecords[pair.first].activeVariants;
+            uint64_t& result = mMaterialRecords[pair.first].activeVariants;
             mQueryCallback(pair.second.userdata, &result);
         }
     }
